Consolidate ActionList stories / knobs FTW
